requirements.txt
# Django và extensions
Django==5.0.1
django-ninja==1.1.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# PostgreSQL
psycopg2-binary==2.9.9

# Web scraping
playwright==1.41.1

# Async support
asgiref==3.7.2

# Utilities
python-dateutil==2.8.2

# Development
ipython==8.20.0
